[refactor] FieldSelection & FieldMapper인터페이스를 infra-database-query 모듈로 이동 #72
