+++ shared/infra-database-query/src/main/kotlin/dev/codeswamp/databasequery/FieldMapper.kt
@@ -1,4 +1,4 @@
-package dev.codeswamp.articlequery.application.service.mapper
+package dev.codeswamp.databasequery
 
 interface FieldMapper {
     fun map(fields: Set<String>): Set<String>
